Unify SubnetSpec type definition
